+++ core/src/HIP/Kokkos_HIP_Team.hpp
@@ -327,9 +327,10 @@ class HIPTeamMember {
   // Private for the driver
 
   KOKKOS_INLINE_FUNCTION
-  HIPTeamMember(void* shared, const int shared_begin, const int shared_size,
-                void* scratch_level_1_ptr, const int scratch_level_1_size,
-                const int arg_league_rank, const int arg_league_size)
+  HIPTeamMember(void* shared, const size_t shared_begin,
+                const size_t shared_size, void* scratch_level_1_ptr,
+                const size_t scratch_level_1_size, const int arg_league_rank,
+                const size_t arg_league_size)
       : m_team_reduce(shared),
         m_team_shared(((char*)shared) + shared_begin, shared_size,
                       scratch_level_1_ptr, scratch_level_1_size),
