+++ core/src/OpenMPTarget/Kokkos_OpenMPTarget_Parallel.hpp
@@ -749,7 +749,7 @@ class ParallelFor<FunctorType, Kokkos::TeamPolicy<Properties...>,
 
   const FunctorType m_functor;
   const Policy m_policy;
-  const int m_shmem_size;
+  const size_t m_shmem_size;
 
  public:
   void execute() const {
@@ -1200,7 +1200,7 @@ class ParallelReduce<FunctorType, Kokkos::TeamPolicy<Properties...>,
   const Policy m_policy;
   const ReducerType m_reducer;
   const pointer_type m_result_ptr;
-  const int m_shmem_size;
+  const size_t m_shmem_size;
 
  public:
   void execute() const {
